Replace the final split admin view.
