+++ pycon/settings.py
@@ -92,7 +92,7 @@ USE_L10N = True
 # If you set this to False, Django will not use timezone-aware datetimes.
 USE_TZ = True
 
-TIME_ZONE = 'Europe/Zurich'
+TIME_ZONE = 'Europe/Dublin'
 LANGUAGE_CODE = 'en'
 
 LANGUAGES = (
